Finish Registration Class
